format: extract size/count/time formatting from ui
The size, count, and time formatting is pure, has nothing to do with ratatui,
and is the part most worth testing on its own, but it was buried in the
rendering module. This is a move only; behavior is unchanged.
